+++ ru.taximaxim.codekeeper.core.tests/src/main/java/ru/taximaxim/codekeeper/core/depcies/ChDiffDepciesTest.java
@@ -34,7 +34,9 @@ class ChDiffDepciesTest {
             // test VIEW depcy
             "add_func_func_1_usr_v1",
             // test scenario where user drop column. user choice TABLE
-            "ch_policy_deps_change_table_usr_t"
+            "ch_policy_deps_change_table_usr_t",
+            // test scenario where added USER with ROLE dep
+            "ch_add_user_timon_usr_timon"
     })
     void testDepcy(final String userSelTemplate) throws IOException, InterruptedException {
         TestUtils.testDepcy(userSelTemplate, true, DatabaseType.CH, getClass());