+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/add_ch_role_diff.sql
@@ -1,0 +1,6 @@
+CREATE ROLE shaiHulud;
+
+CREATE ROLE worm
+	IN ldadp;
+
+CREATE ROLE fox;