+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/compare_ch_role_diff.sql
@@ -1,0 +1,5 @@
+MOVE ROLE admin1 TO ldadp;
+
+MOVE ROLE student TO local_directory;
+
+MOVE ROLE tom TO local_directory;