+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/ch_add_user_timon_new.sql
@@ -1,0 +1,4 @@
+CREATE ROLE admin;
+
+CREATE USER timon
+	DEFAULT ROLE admin;