+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_ch_role_diff.sql
@@ -1,0 +1,5 @@
+DROP ROLE shaiHulud;
+
+DROP ROLE worm;
+
+DROP ROLE fox;