+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_ch_role_original.sql
@@ -1,0 +1,7 @@
+CREATE ROLE shaiHulud;
+
+CREATE ROLE worm
+    IN ldadp;
+    
+CREATE ROLE fox
+    IN local_directory;